refactor(test): modify code to use new paths
