+++ e2e/suite_test.go
@@ -25,10 +25,10 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/openclarity/vmclarity/e2e/config"
-	"github.com/openclarity/vmclarity/e2e/testenv"
-	"github.com/openclarity/vmclarity/e2e/testenv/types"
 	"github.com/openclarity/vmclarity/pkg/shared/backendclient"
 	"github.com/openclarity/vmclarity/pkg/shared/uibackendclient"
+	"github.com/openclarity/vmclarity/testenv"
+	"github.com/openclarity/vmclarity/testenv/types"
 	"github.com/openclarity/vmclarity/utils/log"
 )
