+++ testenv/docker/config.go
@@ -18,7 +18,7 @@ package docker
 import (
 	"context"
 
-	envtypes "github.com/openclarity/vmclarity/e2e/testenv/types"
+	envtypes "github.com/openclarity/vmclarity/testenv/types"
 )
 
 type ContainerImages = envtypes.ContainerImages[string]
