+++ testenv/kubernetes/config.go
@@ -22,9 +22,9 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	"github.com/openclarity/vmclarity/e2e/testenv/kubernetes/helm"
-	"github.com/openclarity/vmclarity/e2e/testenv/kubernetes/types"
-	envtypes "github.com/openclarity/vmclarity/e2e/testenv/types"
+	"github.com/openclarity/vmclarity/testenv/kubernetes/helm"
+	"github.com/openclarity/vmclarity/testenv/kubernetes/types"
+	envtypes "github.com/openclarity/vmclarity/testenv/types"
 )
 
 const (
