+++ testenv/kubernetes/helm/config.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 	"path/filepath"
 
-	envtypes "github.com/openclarity/vmclarity/e2e/testenv/types"
+	envtypes "github.com/openclarity/vmclarity/testenv/types"
 )
 
 const (
